Presentation API polyfill with short hello world example
This repository now contains the Presentation API polyfill which was up until
now somewhat hidden in:
https://github.com/webscreens/slidyremote/

The "presentation-api-polyfill.js" file contains the full polyfill. It is the
result of concatenating "core.js" with the different presentation mechanisms
defined in "mechanisms/".

The examples folder contains a first "Hello world" example.

The README still needs to be completed and an overview page should be created.