Added new QR code presentation mechanism
The code reuses the related "trigger" code from the Association API:
https://github.com/mediascape/association

The QR code is displayed as a <div> overlay. As with other invitation-based
mechanisms, the "controlling" side has no way to determine when a receiver
connects and cannot establish a communication channel with it.

(A few other bug fixes included in this commit)