Implemented "isChannelOptional" presentation request flag
The "PresentationRequest" constructor now also takes an options object as
second parameter. Presentation mechanisms receive the URL and the options
object when "getAvailableDisplays" is called.

Mechanisms that cannot establish a communication channel with the second
screen must check the "isChannelOptional" option. If not set (the default),
they should not return any available display.